// src/Common/Enums.ts
export enum DataCenter {
  UK = 'uk',
  RO = 'ro',
  RO2 = 'ro2',
  RS = 'rs',
  RS2 = 'rs2',
}

export enum Environment {
  PRODUCTION = 'production',
  STAGING = 'staging',
  DEVELOPMENT = 'development',
}

export enum Channel {
  Retail = 'Retail',
  Terminal = 'Terminal',
  Web = 'Web',
  Mobile = 'Mobile'
}